refactor(cache): move inline tests to sibling _tests.rs files
Follows the javm-cap::cap_tests.rs convention. Test modules now live
in sibling files declared in the parent module:

- nub-host-common::cache::directory_tests
- nub-host-common::cache::talc_alloc_tests
- nub-host-common::cache::talc_arc_tests
- nub_host_kvm::cache_tests

Cache::pinned visibility widened to pub(crate) so the moved test
module can still inspect the pinned-set size.

No behavior change; same tests run, same results.
